(refactor) Port useVitalsConceptMetadata from common lib to vitals
