--- packages/esm-patient-common-lib/src/useVitalsConceptMetadata.tsx
@@ -1,62 +1,0 @@
-import useSWRImmutable from 'swr/immutable';
-import { openmrsFetch, restBaseUrl } from '@openmrs/esm-framework';
-
-export function useVitalsConceptMetadata() {
-  const customRepresentation =
-    'custom:(setMembers:(uuid,display,hiNormal,hiAbsolute,hiCritical,lowNormal,lowAbsolute,lowCritical,units))';
-
-  const apiUrl = `${restBaseUrl}/concept/?q=VITALS SIGNS&v=${customRepresentation}`;
-
-  const { data, error, isLoading } = useSWRImmutable<{ data: VitalsConceptMetadataResponse }, Error>(
-    apiUrl,
-    openmrsFetch,
-  );
-
-  const conceptMetadata = data?.data?.results[0]?.setMembers;
-
-  const conceptUnits = conceptMetadata?.length
-    ? new Map<string, string>(conceptMetadata.map((concept) => [concept.uuid, concept.units]))
-    : new Map<string, string>([]);
-
-  const conceptRanges = conceptMetadata?.length
-    ? new Map<string, { lowAbsolute: number | null; highAbsolute: number | null }>(
-        conceptMetadata.map((concept) => [
-          concept.uuid,
-          {
-            lowAbsolute: concept.lowAbsolute ?? null,
-            highAbsolute: concept.hiAbsolute ?? null,
-          },
-        ]),
-      )
-    : new Map<string, { lowAbsolute: number | null; highAbsolute: number | null }>([]);
-
-  return {
-    data: conceptUnits,
-    isError: error,
-    isLoading,
-    conceptMetadata,
-    conceptRanges,
-  };
-}
-
-export const withUnit = (label: string, unit: string | null | undefined) => {
-  return `${label} ${unit ? `(${unit})` : ''}`;
-};
-
-export interface ConceptMetadata {
-  uuid: string;
-  display: string;
-  hiNormal: number | null;
-  hiAbsolute: number | null;
-  hiCritical: number | null;
-  lowNormal: number | null;
-  lowAbsolute: number | null;
-  lowCritical: number | null;
-  units: string | null;
-}
-
-interface VitalsConceptMetadataResponse {
-  results: Array<{
-    setMembers: Array<ConceptMetadata>;
-  }>;
-}
