+++ packages/esm-patient-vitals-app/src/common/data.resource.ts
@@ -7,8 +7,9 @@ import {
   openmrsFetch,
   useConfig,
 } from '@openmrs/esm-framework';
-import { type ObsRecord, useVitalsConceptMetadata, type ConceptMetadata } from '@openmrs/esm-patient-common-lib';
+import { type ObsRecord } from '@openmrs/esm-patient-common-lib';
 import { type KeyedMutator } from 'swr';
+import useSWRImmutable from 'swr/immutable';
 import useSWRInfinite from 'swr/infinite';
 import { type ConfigObject } from '../config-schema';
 import type { FHIRSearchBundleResponse, MappedVitals, PatientVitals, VitalsResponse } from './types';
@@ -32,6 +33,66 @@ type VitalsAndBiometricsSwrKey = {
 };
 
 type VitalsFetchResponse = FetchResponse<VitalsResponse>;
+
+export interface ConceptMetadata {
+  uuid: string;
+  display: string;
+  hiNormal: number | null;
+  hiAbsolute: number | null;
+  hiCritical: number | null;
+  lowNormal: number | null;
+  lowAbsolute: number | null;
+  lowCritical: number | null;
+  units: string | null;
+}
+
+interface VitalsConceptMetadataResponse {
+  results: Array<{
+    setMembers: Array<ConceptMetadata>;
+  }>;
+}
+
+export function useVitalsConceptMetadata() {
+  const customRepresentation =
+    'custom:(setMembers:(uuid,display,hiNormal,hiAbsolute,hiCritical,lowNormal,lowAbsolute,lowCritical,units))';
+
+  const apiUrl = `${restBaseUrl}/concept/?q=VITALS SIGNS&v=${customRepresentation}`;
+
+  const { data, error, isLoading } = useSWRImmutable<{ data: VitalsConceptMetadataResponse }, Error>(
+    apiUrl,
+    openmrsFetch,
+  );
+
+  const conceptMetadata = data?.data?.results[0]?.setMembers;
+
+  const conceptUnits = conceptMetadata?.length
+    ? new Map<string, string>(conceptMetadata.map((concept) => [concept.uuid, concept.units]))
+    : new Map<string, string>([]);
+
+  const conceptRanges = conceptMetadata?.length
+    ? new Map<string, { lowAbsolute: number | null; highAbsolute: number | null }>(
+        conceptMetadata.map((concept) => [
+          concept.uuid,
+          {
+            lowAbsolute: concept.lowAbsolute ?? null,
+            highAbsolute: concept.hiAbsolute ?? null,
+          },
+        ]),
+      )
+    : new Map<string, { lowAbsolute: number | null; highAbsolute: number | null }>([]);
+
+  return {
+    data: conceptUnits,
+    error,
+    isLoading,
+    conceptMetadata,
+    conceptRanges,
+  };
+}
+
+export const withUnit = (label: string, unit: string | null | undefined) => {
+  return `${label} ${unit ? `(${unit})` : ''}`;
+};
 
 // We need to track a bound mutator for basically every hook, because there does not appear to be
 // a way to invalidate an SWRInfinite key that works other than using the bound mutator
@@ -77,7 +138,7 @@ export function useVitalsAndBiometrics(patientUuid: string, mode: VitalsAndBiome
     [swrKeyNeedle, mode, conceptUuids],
   );
 
-  const { data, isValidating, setSize, error, size, mutate } = useSWRInfinite<VitalsFetchResponse, Error>(
+  const { data, isLoading, isValidating, setSize, error, size, mutate } = useSWRInfinite<VitalsFetchResponse, Error>(
     getPage,
     handleFetch,
   );
@@ -165,8 +226,8 @@ export function useVitalsAndBiometrics(patientUuid: string, mode: VitalsAndBiome
 
   return {
     data: data ? formattedObs : undefined,
-    isLoading: !data && !error,
-    isError: error,
+    isLoading,
+    error,
     hasMore: data?.length
       ? !!data[data.length - 1].data?.link?.some((link: { relation?: string }) => link.relation === 'next')
       : false,
